scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.1",
    "python-benedict[all]>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
